Truly move on from my Atlassian past by stopping the over-moduleification.
I probably didn't do a great job of it 2 years ago, but the original idea was to be able to have a choice between

* The api
* The core runner
* The before/after block defaults

honestly. It was all a bit of YAGNI. This is taking it back out. Simplicity is good. The API stays out for now, but may come in to the core module in the future. For now though - i like the idea of the one way dependency. Ensures that implementation doesn't escape into the API